eyihcn

+++ reliable-message/provider-message-service-api/src/main/java/com/fuli/cloud/message/provider/aspect/ConsumerAroundHandlerAspect.java
@@ -13,7 +13,6 @@ import com.fuli.cloud.message.provider.exceptions.FuliMqMessageBizException;
 import com.fuli.cloud.message.provider.model.domain.FuliMqMessage;
 import com.fuli.cloud.message.provider.model.enums.ConsumeConcurrentlyStatusEnum;
 import com.fuli.cloud.message.provider.service.feign.FuliMqMessageFeignApi;
-import com.fuli.cloud.message.provider.util.MessageBizResultUtils;
 
 import eyihcn.common.core.enums.ErrorCodeEnum;
 import lombok.extern.slf4j.Slf4j;
@@ -63,30 +62,25 @@ public class ConsumerAroundHandlerAspect {
 		if (!(args[0] instanceof FuliMqMessage)) {
 			throw new FuliMqMessageBizException(ErrorCodeEnum.MESSAGE_SDK10050011, methodName);
 		}
-		FuliMqMessage fuliMqMessage = null;
-		try {
-			fuliMqMessage = (FuliMqMessage) args[0];
-		} catch (Exception e) {
-			throw new FuliMqMessageBizException(ErrorCodeEnum.MESSAGE_SDK10050011);
-		}
+		FuliMqMessage fuliMqMessage = (FuliMqMessage) args[0];
 		final String messageId = fuliMqMessage.getId();
 		// 业务执行结果
-		joinPoint.proceed();
-		ConsumeConcurrentlyStatusEnum consumeResult = MessageBizResultUtils.getConsumeResult();
+
+		Object proceed = joinPoint.proceed();
+		if (!(proceed instanceof ConsumeConcurrentlyStatusEnum)) {
+			throw new RuntimeException("请返回业务消费消息结果枚举类型");
+		}
+		ConsumeConcurrentlyStatusEnum consumeResult = (ConsumeConcurrentlyStatusEnum) proceed;
 		log.info("消费者，目标方法={}, 业务执行结果, result={}", methodName, consumeResult);
 		try {
 			if (ConsumeConcurrentlyStatusEnum.CONSUME_SUCCESS == consumeResult) {
 				// 消息成功消费，删除消息
-				// TODO 逻辑删除
 				log.info("被动方业务执行成功，消费者切面开始异步删除消息，消息id为[{}]", messageId);
 				taskExecutor.execute(() -> fuliMqMessageFeignApi.deleteMessageByMessageId(messageId));
 			}
 		} catch (Exception e) {
-			log.error("消费可靠消息出现异常, 目标方法[{}], 异常={},总耗时={}ms", methodName, e.getMessage(),
-					System.currentTimeMillis() - startTime);
 			throw e;
 		} finally {
-			MessageBizResultUtils.clearConsumeResult();
 			log.info("消费可靠消息 目标方法[{}], 总耗时={}ms", methodName, System.currentTimeMillis() - startTime);
 		}
 	}
